+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - [spiral/queue] Added new option `headers` in the `Spiral\Queue\Options` and new interface 
     `Spiral\Queue\ExtendedOptionsInterface`.
   - [spiral/events] Added event interceptors.
+  - [spiral/core] Added container instance to callback function parameters in `Spiral\Core\Container` and 
+    `Spiral\Core\ContainerScope`.       
 
 ## v3.1.0 - 2022-09-29
 - **Other Features**